+++ public/config.js
@@ -1,3 +1,4 @@
+// "kimaiAPI" : "http://localhost/api/",
 var config = {
     "kimaiAPI" : "https://demo-stable.kimai.org/api/",
     "ignoreActivities" : ["Vacation","Public Holiday"],
@@ -6,5 +7,6 @@ var config = {
     "overtime" : false,
     "summary15mins" : false,
     "breakChecks" : false,
-    "bundleApiUpdates" : false,
+    "bundleApiUpdates" : true,
+    "loginToken" : true,
 }