+++ src/api/kimaiClientPostGeneric.js
@@ -5,8 +5,12 @@ export async function kimaiClientPostGeneric(method, config, data, type = "POST"
         req.open(type, config.kimaiAPI + method, true);
         req.responseType = "json";
         req.setRequestHeader('Content-Type', 'application/json');
-        req.setRequestHeader('X-AUTH-USER', config.username);
-        req.setRequestHeader('X-AUTH-TOKEN', config.password);
+        if (config.loginToken !== true) {
+            req.setRequestHeader('X-AUTH-USER', config.username);
+            req.setRequestHeader('X-AUTH-TOKEN', config.password);
+        } else {
+            req.setRequestHeader('Authorization', 'Bearer ' + config.token);
+        }
         req.setRequestHeader('Cache-Control', 'no-cache, no-store, max-age=0');
 
         req.onload = () => resolve(req.response);