+++ src/features/views/Info.js
@@ -12,7 +12,7 @@ function Info (props) {
     const kimaiGeneralStatusError = useSelector((state) => state.kimaiData.error);
 
     if (
-        !(config.username && config.password) ||
+        !((config.username && config.password) || (config.loginToken === true && config.token)) ||
         (kimaiGeneralStatusError &&
             kimaiGeneralStatusError === 'Invalid credentials')
     ) {